refactor: detailed-schedule 리팩토링
