+++ src/entities/detailed-schedule/__stories__/detailed-schedule-card-content.stories.tsx
