+++ src/entities/detailed-schedule/__stories__/detailed-schedule-color-badge.stories.tsx
