+++ src/entities/detailed-schedule/__stories__/detailed-schedule-list-card.stories.tsx
