+++ src/entities/detailed-schedule/ui/detailed-schedule-card-content.tsx
