+++ src/entities/detailed-schedule/ui/detailed-schedule-date.tsx
