+++ src/entities/detailed-schedule/ui/index.ts
@@ -1,0 +1,5 @@
+export { default as DetailedScheduleCard } from './detailed-schedule-card'
+export { default as DetailedScheduleCardContent } from './detailed-schedule-card-content'
+export { default as DetailedScheduleColorBadge } from './detailed-schedule-color-badge'
+export { default as DetailedScheduleDate } from './detailed-schedule-date'
+export { default as DetailedScheduleListCard } from './detailed-schedule-list-card'
