+++ src/entities/schedule/api/detailed-schedule.API.ts
@@ -1,6 +1,6 @@
 import axiosInstance from '@/shared/api/axios-instance'
 
-import type { Schedule } from '@/entities/schedule/models'
+import type { Schedule } from '@/entities/schedule/lib'
 
 export async function getDetailedSchedule(scheduleId: string): Promise<Schedule> {
   const res = await axiosInstance(`/schedules/${scheduleId}`)
