+++ src/entities/schedule/api/get-friend-schedule.API.ts
@@ -1,6 +1,6 @@
 import axiosInstance from '@/shared/api/axios-instance'
 
-import type { Schedule } from '@/entities/schedule/models'
+import type { Schedule } from '@/entities/schedule/lib'
 
 interface Props {
   friendId: string
