+++ src/entities/schedule/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './use-my-schedules-by-month'
 export * from './use-delete-schedule'
 export * from './use-friend-schedule-by-userId'
 export * from './use-share-schedule-by-userId'
+export * from './use-month-store'
