+++ src/entities/schedule/hooks/use-detailed-schedule.ts
@@ -1,7 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 
 import { getDetailedSchedule } from '../api'
-import { scheduleQueryKeys } from '../models'
+import { scheduleQueryKeys } from '../lib'
 
 export function useDetailedSchedule(scheduleId: string) {
   const { data } = useSuspenseQuery({
