+++ src/entities/schedule/hooks/use-friend-schedule-by-userId.ts
@@ -1,7 +1,7 @@
 import { useQueries } from '@tanstack/react-query'
 import { format, startOfMonth, endOfMonth } from 'date-fns'
 
-import { groupByDate, scheduleQueryKeys, type Schedule } from '@/entities/schedule/models'
+import { groupByDate, scheduleQueryKeys, type Schedule } from '@/entities/schedule/lib'
 
 import { getFriendSchedule } from '../api'
