+++ src/entities/schedule/lib/get-group-by-date.ts
