+++ src/entities/schedule/lib/index.ts
@@ -1,5 +1,5 @@
 export * from './schedule.types'
 export * from './get-group-by-date'
 export * from './schedule.query'
-export * from './use-month-store'
 export * from './month-to-range'
+export * from './formatted-date'
