+++ src/entities/schedule/lib/schedule.query.ts
