+++ src/entities/schedule/lib/schedule.types.ts
