+++ src/features/calendar/ui/__stories__/render-schedule-badge.stories.tsx
@@ -1,6 +1,6 @@
 import { RenderScheduleBadges } from '..'
 
-import type { Schedule } from '@/entities/schedule/models'
+import type { Schedule } from '@/entities/schedule/lib'
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
 const meta: Meta<typeof RenderScheduleBadges> = {
