+++ src/features/detailed-schedule/ui/friend-detailed-schedule-list.tsx
@@ -1,28 +1,32 @@
 import { useState } from 'react'
 import { useNavigate, useParams } from 'react-router'
 
-import { useShareScheduleByUserId } from '@/entities/schedule/hooks'
+import { DetailedScheduleListCard } from '@/entities/detailed-schedule/ui'
+import { useFriendSchedulesByuserId } from '@/entities/schedule/hooks'
+import { useFriendMonths } from '@/entities/schedule/lib'
 import { formatScheduleTime } from '@/entities/utils/format-schedule-time'
 import { PrivateScheduleModal } from '@/features/calendar/ui'
-import { DetailedScheduleListCard } from '@/features/detailed-schedule/ui'
-import { IconInvite } from '@/shared/assets/icons'
+import { IconInvite } from '@/shared/assets'
 import { FloatButton } from '@/shared/ui'
 
-export default function ShareDetailedScheduleList() {
+export default function FriendDetailedScheduleList() {
   const navigate = useNavigate()
-  const { date, userId } = useParams() // ex: '2025-08-01'
-  const { scheduleMap } = useShareScheduleByUserId()
+  const { date, friendId } = useParams() // ex: '2025-08-01'
+
+  const { months } = useFriendMonths(friendId!)
+
+  const { scheduleMap } = useFriendSchedulesByuserId({ months, friendId: friendId! })
+
   const list = date ? (scheduleMap[date] ?? []) : []
+
   const [isModalOpen, setIsModalOpen] = useState(false)
 
   const onClick = (scheduleId: number) => {
-    navigate(`/share/${userId}/detailed-schedule/date/${date}/schedules/${scheduleId}`)
+    navigate(`/friends/${friendId}/calendar/detailed-schedule/date/${date}/schedules/${scheduleId}`)
   }
-
   const goToCreateAppointment = () => {
-    navigate(`/share/${userId}/appointment/create?date=${date}`)
+    navigate(`/friends/${friendId}/calendar/appointment/create?date=${date}`)
   }
-
   return (
     <div>
       {list.length === 0 ? (
@@ -30,12 +34,13 @@ export default function ShareDetailedScheduleList() {
       ) : (
         <div className="flex flex-col gap-4">
           {list.map((card) => {
+            const title = card.isVisible ? card.title : '비공개일정'
             const handleClick = card.isVisible ? () => onClick(card.scheduleId) : () => setIsModalOpen(true)
             return (
               <DetailedScheduleListCard
                 onCardClick={handleClick}
                 key={`${card.scheduleId}-${card.startAt}`}
-                title={card.title}
+                title={title}
                 time={formatScheduleTime(card)}
                 badgeColor={card.color}
                 isVisible={card.isVisible}
