+++ src/features/detailed-schedule/ui/index.ts
@@ -1,6 +1,7 @@
-export { default as DetailedScheduleCard } from './detailed-schedule-card'
-export { default as DetailedScheduleCardContent } from './detailed-schedule-card-content'
-export { default as DetailedScheduleColorBadge } from './detailed-schedule-color-badge'
-export { default as DetailedScheduleDate } from './detailed-schedule-date'
 export { default as DetailedScheduleHeader } from './detailed-schedule-header'
-export { default as DetailedScheduleListCard } from './detailed-schedule-list-card'
+export { default as FriendDetailedScheduleList } from './friend-detailed-schedule-list'
+export { default as FriendDetailedSchedule } from './friend-detailed-schedule'
+export { default as MyDetailedScheduleList } from './my-detailed-schedule-list'
+export { default as MyDetailedSchedule } from './my-detailed-schedule'
+export { default as ShareDetailedScheduleList } from './share-detailed-schedule-list'
+export { default as ShareDetailedSchedule } from './share-detailed-schedule'
