+++ src/features/detailed-schedule/ui/my-detailed-schedule.tsx
@@ -1,12 +1,13 @@
 import { useParams } from 'react-router'
 
+import { DetailedScheduleCard } from '@/entities/detailed-schedule/ui'
 import { useDetailedSchedule } from '@/entities/schedule/hooks'
-import { getRepeatText } from '@/entities/utils/format-repeat-text'
-import DetailedScheduleCard from '@/shared/ui/detailed-schedule/detailed-schedule-card'
+import { getRepeatText } from '@/entities/utils'
 
-export default function FriendDetailedSchedule() {
+export default function MyDetailedSchedule() {
   const { scheduleId } = useParams()
   const data = useDetailedSchedule(scheduleId!)
+
   const repeatText = getRepeatText({
     repeatRule: data?.repeatRule ?? '',
     repeatType: data?.repeatType ?? null,
@@ -20,7 +21,7 @@ export default function FriendDetailedSchedule() {
       title={data.title}
       startDate={data.startAt}
       endDate={data.endAt}
-      repeat={repeatText ?? ''}
+      repeat={repeatText}
       visible={data.isVisible}
       content={data.content}
     />
