+++ src/features/detailed-schedule/ui/share-detailed-schedule-list.tsx
@@ -1,32 +1,28 @@
 import { useState } from 'react'
 import { useNavigate, useParams } from 'react-router'
 
-import { useFriendSchedulesByuserId } from '@/entities/schedule/hooks'
-import { useFriendMonths } from '@/entities/schedule/models'
+import { DetailedScheduleListCard } from '@/entities/detailed-schedule/ui'
+import { useShareScheduleByUserId } from '@/entities/schedule/hooks'
 import { formatScheduleTime } from '@/entities/utils/format-schedule-time'
 import { PrivateScheduleModal } from '@/features/calendar/ui'
-import { DetailedScheduleListCard } from '@/features/detailed-schedule/ui'
-import { IconInvite } from '@/shared/assets'
+import { IconInvite } from '@/shared/assets/icons'
 import { FloatButton } from '@/shared/ui'
 
-export default function FriendDetailedScheduleList() {
+export default function ShareDetailedScheduleList() {
   const navigate = useNavigate()
-  const { date, friendId } = useParams() // ex: '2025-08-01'
-
-  const { months } = useFriendMonths(friendId!)
-
-  const { scheduleMap } = useFriendSchedulesByuserId({ months, friendId: friendId! })
-
+  const { date, userId } = useParams() // ex: '2025-08-01'
+  const { scheduleMap } = useShareScheduleByUserId()
   const list = date ? (scheduleMap[date] ?? []) : []
-
   const [isModalOpen, setIsModalOpen] = useState(false)
 
   const onClick = (scheduleId: number) => {
-    navigate(`/friends/${friendId}/calendar/detailed-schedule/date/${date}/schedules/${scheduleId}`)
+    navigate(`/share/${userId}/detailed-schedule/date/${date}/schedules/${scheduleId}`)
   }
+
   const goToCreateAppointment = () => {
-    navigate(`/friends/${friendId}/calendar/appointment/create?date=${date}`)
+    navigate(`/share/${userId}/appointment/create?date=${date}`)
   }
+
   return (
     <div>
       {list.length === 0 ? (
@@ -34,13 +30,12 @@ export default function FriendDetailedScheduleList() {
       ) : (
         <div className="flex flex-col gap-4">
           {list.map((card) => {
-            const title = card.isVisible ? card.title : '비공개일정'
             const handleClick = card.isVisible ? () => onClick(card.scheduleId) : () => setIsModalOpen(true)
             return (
               <DetailedScheduleListCard
                 onCardClick={handleClick}
                 key={`${card.scheduleId}-${card.startAt}`}
-                title={title}
+                title={card.title}
                 time={formatScheduleTime(card)}
                 badgeColor={card.color}
                 isVisible={card.isVisible}
