--- src/features/friend-schedule/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as FriendDetailedScheduleList } from './friend-detailed-schedule-list'
-export { default as FriendDetailedSchedule } from './friend-detailed-schedule'
