+++ src/features/my-calendar/ui/index.ts
@@ -1,6 +1,2 @@
-export { default as DeleteConfirmModal } from './delete-confirm-modal'
-export { default as MyDetailedScheduleList } from './my-detailed-schedule-list'
-export { default as MyDetailedSchedule } from './my-detailed-schedule'
 export { default as ShareCalendarBottomSheet } from './share-calendar-bottom-sheet'
-export { default as EditSchduleDropDown } from './edit-schedule-dropdown'
 export { default as SwitchModal } from './switch-modal'
