--- src/features/share-schedule/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as ShareDetailedSchedule } from './share-detailed-schedule'
-export { default as ShareDetailedScheduleList } from './share-detailed-schedule-list'
